Move buildOperationList to renderer.ts.
